+++ packages/expo-router/src/__tests__/tabs.test.ios.tsx
@@ -134,14 +134,21 @@ it('has correct routeInfo when switching tabs as a nested navigator - using pres
     }
   );
 
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
+  expect(layoutCalls).toHaveBeenCalledWith(['(tabs)']);
+
+  expect(indexCalls).toHaveBeenCalledTimes(1);
+  expect(indexCalls).toHaveBeenCalledWith(['(tabs)']);
+
+  expect(exploreCalls).toHaveBeenCalledTimes(0);
+
+  jest.clearAllMocks();
   fireEvent.press(screen.getByLabelText('explore, tab, 2 of 2'));
 
-  expect(layoutCalls).toHaveBeenCalledTimes(2);
-  expect(layoutCalls).toHaveBeenNthCalledWith(1, ['(tabs)']);
-  expect(layoutCalls).toHaveBeenNthCalledWith(2, ['(tabs)', 'explore']);
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
+  expect(layoutCalls).toHaveBeenCalledWith(['(tabs)', 'explore']);
 
-  expect(indexCalls).toHaveBeenCalledTimes(2);
-  expect(indexCalls).toHaveBeenCalledWith(['(tabs)', 'explore']);
+  expect(indexCalls).toHaveBeenCalledTimes(1);
   expect(indexCalls).toHaveBeenCalledWith(['(tabs)', 'explore']);
 
   expect(exploreCalls).toHaveBeenCalledTimes(1);
@@ -150,9 +157,8 @@ it('has correct routeInfo when switching tabs as a nested navigator - using pres
   jest.clearAllMocks();
   fireEvent.press(screen.getByLabelText('index, tab, 1 of 2'));
 
-  expect(layoutCalls).toHaveBeenCalledTimes(2);
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
   expect(layoutCalls).toHaveBeenNthCalledWith(1, ['(tabs)']);
-  expect(layoutCalls).toHaveBeenNthCalledWith(2, ['(tabs)']);
 
   expect(indexCalls).toHaveBeenCalledTimes(1);
   expect(indexCalls).toHaveBeenCalledWith(['(tabs)']);
@@ -163,16 +169,91 @@ it('has correct routeInfo when switching tabs as a nested navigator - using pres
   jest.clearAllMocks();
   fireEvent.press(screen.getByLabelText('explore, tab, 2 of 2'));
 
-  expect(layoutCalls).toHaveBeenCalledTimes(2);
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
   expect(layoutCalls).toHaveBeenNthCalledWith(1, ['(tabs)', 'explore']);
-  expect(layoutCalls).toHaveBeenNthCalledWith(2, ['(tabs)', 'explore']);
 
-  expect(indexCalls).toHaveBeenCalledTimes(2);
-  expect(indexCalls).toHaveBeenCalledWith(['(tabs)', 'explore']);
+  expect(indexCalls).toHaveBeenCalledTimes(1);
   expect(indexCalls).toHaveBeenCalledWith(['(tabs)', 'explore']);
 
   expect(exploreCalls).toHaveBeenCalledTimes(1);
   expect(exploreCalls).toHaveBeenCalledWith(['(tabs)', 'explore']);
+});
+
+it('has correct routeInfo when switching tabs using press', () => {
+  /**
+   * This test exists because there are inconsistencies when using press vs the API.
+   * This is due to how React Navigation fires events on press (inconsistent) vs API calls (consistent).
+   */
+  const layoutCalls = jest.fn();
+  const indexCalls = jest.fn();
+  const exploreCalls = jest.fn();
+
+  /**
+   * In this instance, React Navigation fires the state update before the screen is rendered.
+   */
+  renderRouter(
+    {
+      _layout: function Layout() {
+        layoutCalls(useSegments());
+        return <Tabs />;
+      },
+      index: function Index() {
+        indexCalls(useSegments());
+        return <Text testID="index">Index</Text>;
+      },
+      explore: function Explore() {
+        exploreCalls(useSegments());
+        return <Text testID="explore">Explore</Text>;
+      },
+    },
+    {
+      initialUrl: '/explore',
+    }
+  );
+
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
+  expect(layoutCalls).toHaveBeenCalledWith(['explore']);
+
+  expect(exploreCalls).toHaveBeenCalledTimes(1);
+  expect(exploreCalls).toHaveBeenCalledWith(['explore']);
+
+  expect(indexCalls).toHaveBeenCalledTimes(0);
+
+  jest.clearAllMocks();
+  fireEvent.press(screen.getByLabelText('index, tab, 1 of 4'));
+
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
+  expect(layoutCalls).toHaveBeenNthCalledWith(1, []);
+
+  expect(indexCalls).toHaveBeenCalledTimes(1);
+  expect(indexCalls).toHaveBeenCalledWith([]);
+
+  expect(exploreCalls).toHaveBeenCalledTimes(1);
+  expect(exploreCalls).toHaveBeenCalledWith([]);
+
+  jest.clearAllMocks();
+  fireEvent.press(screen.getByLabelText('explore, tab, 2 of 4'));
+
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
+  expect(layoutCalls).toHaveBeenCalledWith(['explore']);
+
+  expect(indexCalls).toHaveBeenCalledTimes(1);
+  expect(indexCalls).toHaveBeenCalledWith(['explore']);
+
+  expect(exploreCalls).toHaveBeenCalledTimes(1);
+  expect(exploreCalls).toHaveBeenCalledWith(['explore']);
+
+  jest.clearAllMocks();
+  fireEvent.press(screen.getByLabelText('index, tab, 1 of 4'));
+
+  expect(layoutCalls).toHaveBeenCalledTimes(1);
+  expect(layoutCalls).toHaveBeenNthCalledWith(1, []);
+
+  expect(indexCalls).toHaveBeenCalledTimes(1);
+  expect(indexCalls).toHaveBeenCalledWith([]);
+
+  expect(exploreCalls).toHaveBeenCalledTimes(1);
+  expect(exploreCalls).toHaveBeenCalledWith([]);
 });
 
 it('can push screens', () => {
@@ -344,4 +425,43 @@ it('can use replace navigation with history backBehavior', () => {
   act(() => router.back());
 
   expect(screen.getByTestId('one')).toBeVisible();
+});
+
+it('does not re-render when navigating to different tab', () => {
+  const onOneRender = jest.fn();
+  const onTwoRender = jest.fn();
+  renderRouter(
+    {
+      _layout: () => <Tabs />,
+      one: function One() {
+        onOneRender();
+        return <Text testID="one">One</Text>;
+      },
+      two: function Two() {
+        onTwoRender();
+        return <Text testID="two">Two</Text>;
+      },
+    },
+    {
+      initialUrl: '/one',
+    }
+  );
+
+  expect(screen.getByTestId('one')).toBeVisible();
+  expect(onOneRender).toHaveBeenCalledTimes(1);
+  expect(onTwoRender).toHaveBeenCalledTimes(0);
+
+  jest.clearAllMocks();
+  act(() => router.push('/two'));
+
+  expect(screen.getByTestId('two')).toBeVisible();
+  expect(onOneRender).toHaveBeenCalledTimes(0);
+  expect(onTwoRender).toHaveBeenCalledTimes(1);
+
+  jest.clearAllMocks();
+  act(() => router.push('/one'));
+
+  expect(screen.getByTestId('one')).toBeVisible();
+  expect(onOneRender).toHaveBeenCalledTimes(1);
+  expect(onTwoRender).toHaveBeenCalledTimes(0);
 });